requirements.txt
astroid==2.13.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
fonttools==4.38.0
isort==5.11.4
joblib==1.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.6.3
mccabe==0.7.0
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
platformdirs==2.6.2
psutil==5.9.4
pylint==2.15.10
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
